+++ mods/Mining/mod/prototypes/recipe/intermediate_recipe.lua
@@ -1,0 +1,121 @@
+require("__RsUtility__.scripts.prototypes")
+
+local bronze_nails = {
+  name = "rs-bronze-nails",
+  category = "crafting",
+  ingredients = {
+    {"rs-bronze-bar", 1}
+  },
+  results = {
+    {"rs-bronze-nails", 5}
+  },
+  energy_required = 1.2
+}
+create_recipe(bronze_nails)
+
+local iron_nails = {
+  name = "rs-iron-nails",
+  category = "crafting",
+  ingredients = {
+    {"rs-iron-bar", 1}
+  },
+  results = {
+    {"rs-iron-nails", 5}
+  },
+  energy_required = 1.2
+}
+create_recipe(iron_nails)
+
+local steel_nails = {
+  name = "rs-steel-nails",
+  category = "crafting",
+  ingredients = {
+    {"rs-steel-bar", 1}
+  },
+  results = {
+    {"rs-steel-nails", 5}
+  },
+  energy_required = 1.2
+}
+create_recipe(steel_nails)
+
+local mithril_nails = {
+  name = "rs-mithril-nails",
+  category = "crafting",
+  ingredients = {
+    {"rs-mithril-bar", 1}
+  },
+  results = {
+    {"rs-mithril-nails", 5}
+  },
+  energy_required = 1.2
+}
+create_recipe(mithril_nails)
+
+local adamant_nails = {
+  name = "rs-adamant-nails",
+  category = "crafting",
+  ingredients = {
+    {"rs-adamant-bar", 1}
+  },
+  results = {
+    {"rs-adamant-nails", 5}
+  },
+  energy_required = 1.2
+}
+create_recipe(adamant_nails)
+
+local rune_nails = {
+  name = "rs-rune-nails",
+  category = "crafting",
+  ingredients = {
+    {"rs-rune-bar", 1}
+  },
+  results = {
+    {"rs-rune-nails", 5}
+  },
+  energy_required = 1.2
+}
+create_recipe(rune_nails)
+
+
+local bronze_cable = {
+  name = "rs-bronze-cable",
+  category = "crafting",
+  ingredients = {
+    {"rs-bronze-bar", 1}
+  },
+  results = {
+    {"rs-bronze-cable", 2}
+  },
+  energy_required = 0.6,
+  expensive = "nil",
+  normal = "nil"
+}
+create_recipe(bronze_cable)
+
+local adamant_cable = {
+  name = "rs-adamant-cable",
+  category = "crafting",
+  ingredients = {
+    {"rs-adamant-bar", 1}
+  },
+  results = {
+    {"rs-adamant-cable", 2}
+  },
+  energy_required = 0.6
+}
+create_recipe(adamant_cable)
+
+local orikalkum_cable = {
+  name = "rs-orikalkum-cable",
+  category = "crafting",
+  ingredients = {
+    {"rs-orikalkum-bar", 1}
+  },
+  results = {
+    {"rs-orikalkum-cable", 2}
+  },
+  energy_required = 0.6
+}
+create_recipe(orikalkum_cable)