+++ mods/Mining/mod/prototypes/recipe/pickaxe_recipe.lua
@@ -1,0 +1,131 @@
+require("__RsUtility__.scripts.prototypes")
+
+local bronze_pickaxe = {
+  name = "rs-bronze-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-bronze-bar", 2}
+  },
+  results = {
+    {"rs-bronze-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(bronze_pickaxe)
+
+local iron_pickaxe = {
+  name = "rs-iron-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-iron-bar", 2}
+  },
+  results = {
+    {"rs-iron-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(iron_pickaxe)
+
+local steel_pickaxe = {
+  name = "rs-steel-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-steel-bar", 2}
+  },
+  results = {
+    {"rs-steel-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(steel_pickaxe)
+
+local mithril_pickaxe = {
+  name = "rs-mithril-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-mithril-bar", 2}
+  },
+  results = {
+    {"rs-mithril-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(mithril_pickaxe)
+
+local adamant_pickaxe = {
+  name = "rs-adamant-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-adamant-bar", 2}
+  },
+  results = {
+    {"rs-adamant-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(adamant_pickaxe)
+
+local rune_pickaxe = {
+  name = "rs-rune-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-rune-bar", 2}
+  },
+  results = {
+    {"rs-rune-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(rune_pickaxe)
+
+local orikalkum_pickaxe = {
+  name = "rs-orikalkum-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-orikalkum-bar", 2}
+  },
+  results = {
+    {"rs-orikalkum-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(orikalkum_pickaxe)
+
+local necronium_pickaxe = {
+  name = "rs-necronium-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-necronium-bar", 2}
+  },
+  results = {
+    {"rs-necronium-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(necronium_pickaxe)
+
+local bane_pickaxe = {
+  name = "rs-bane-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-bane-bar", 2}
+  },
+  results = {
+    {"rs-bane-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(bane_pickaxe)
+
+local elder_rune_pickaxe = {
+  name = "rs-elder-rune-pickaxe",
+  category = "crafting",
+  ingredients = {
+    {"rs-elder-rune-bar", 2}
+  },
+  results = {
+    {"rs-elder-rune-pickaxe", 1}
+  },
+  energy_required = 2.4
+}
+create_recipe(elder_rune_pickaxe)